+++ remy/global_injector.go
@@ -1,34 +1,32 @@
 package remy
 
-import "sync"
-
-var (
-	__globalInjector Injector
-	globalMutex      sync.RWMutex
+import (
+	"github.com/wrapped-owls/goremy-di/remy/internal/utils"
 )
 
-// globalInjector retrieves the global injector or generate a new one if it doesn't exist.
-//
-// This is made to prevent loading it unnecessarily into the memory
-func globalInjector() Injector {
-	globalMutex.RLock()
-	if __globalInjector != nil {
-		defer globalMutex.RUnlock()
-		return __globalInjector
+var (
+	__globalInjector = NewInjector()
+)
+
+// mustInjector receives an injector instance, and then returns it if exists or the global injector if it doesn't exist.
+func mustInjector(ij Injector) Injector {
+	if ij != nil {
+		return ij
 	}
-	globalMutex.RUnlock()
+	return __globalInjector
+}
 
-	globalMutex.Lock()
-	defer globalMutex.Unlock()
-
-	// Checks again if no other goroutine has initialized the injector
-	if __globalInjector == nil {
-		__globalInjector = NewInjector()
+func mustRetriever(retriever DependencyRetriever) DependencyRetriever {
+	if retriever != nil {
+		return retriever
 	}
 	return __globalInjector
 }
 
 // SetGlobalInjector receives a custom injector and saves it to be used as a global injector
 func SetGlobalInjector(i Injector) {
+	if i == nil {
+		panic(utils.ErrOverrideInRuntime)
+	}
 	__globalInjector = i
 }