+++ remy/internal/injector/standard_injector.go
@@ -14,8 +14,8 @@ type (
 	}
 )
 
-func New(canOverride bool, reflectOpts types.ReflectionOptions, parent ...types.Injector) *StdInjector {
-	var parentInjector types.Injector
+func New(canOverride bool, reflectOpts types.ReflectionOptions, parent ...types.DependencyRetriever) *StdInjector {
+	var parentInjector types.DependencyRetriever
 	if len(parent) > 0 {
 		parentInjector = parent[0]
 	}
@@ -36,7 +36,11 @@ func (s *StdInjector) SubInjector(overrides ...bool) types.Injector {
 	return New(canOverride, s.reflectOpts, s)
 }
 
-func (s StdInjector) ReflectOpts() types.ReflectionOptions {
+func (s *StdInjector) WrapRetriever() types.Injector {
+	return nil
+}
+
+func (s *StdInjector) ReflectOpts() types.ReflectionOptions {
 	return s.reflectOpts
 }
 
@@ -54,7 +58,7 @@ func (s *StdInjector) BindNamed(bType types.BindKey, name string, value any) err
 	return nil
 }
 
-func (s StdInjector) Get(key types.BindKey) (result any, err error) {
+func (s *StdInjector) Get(key types.BindKey) (result any, err error) {
 	if result, err = s.cacheStorage.Get(key); err != nil && s.parentInjector != nil {
 		result, err = s.parentInjector.Get(key)
 		if err != nil {
@@ -64,7 +68,7 @@ func (s StdInjector) Get(key types.BindKey) (result any, err error) {
 	return
 }
 
-func (s StdInjector) GetNamed(bType types.BindKey, name string) (result any, err error) {
+func (s *StdInjector) GetNamed(bType types.BindKey, name string) (result any, err error) {
 	if result, err = s.cacheStorage.GetNamed(bType, name); err != nil && s.parentInjector != nil {
 		result, err = s.parentInjector.GetNamed(bType, name)
 		if err != nil {