+++ remy/internal/utils/errors.go
@@ -7,4 +7,6 @@ var (
 	ErrImpossibleIdentifyType       = errors.New("impossible to identify the given type")
 	ErrElementNotRegistered         = errors.New("element with given key is not registered")
 	ErrNoElementFoundInsideOrParent = errors.New("no element found on the given injector or any of it's parents")
+	ErrCycleDependencyDetected      = errors.New("cycle dependency detected, check for it")
+	ErrOverrideInRuntime            = errors.New("a process tried to override a value during runtime")
 )