+++ remy/internal/utils/generics_test.go
@@ -1,8 +1,8 @@
 package utils
 
-import (
-	"testing"
-)
+import "testing"
+
+const typeNameErr = "type names was the same, when it should be different, because of different pointer"
 
 func TestTypeName__DifferPointerFromInterface(t *testing.T) {
 	type testInterface interface {
@@ -12,8 +12,12 @@ func TestTypeName__DifferPointerFromInterface(t *testing.T) {
 	for _, generifyInterface := range [...]bool{true, false} {
 		interfaceTypeResult := TypeName[testInterface](generifyInterface)
 		pointerTypeResult := TypeName[*testInterface](generifyInterface)
+		doublePointerTypeResult := TypeName[**testInterface](generifyInterface)
 		if interfaceTypeResult == pointerTypeResult {
-			t.Error("type names was the same, when it should be different, because of different pointer")
+			t.Error(typeNameErr)
+		}
+		if doublePointerTypeResult == pointerTypeResult {
+			t.Error(typeNameErr)
 		}
 	}
 }
@@ -26,8 +30,12 @@ func TestTypeNameByReflect__DifferPointerFromInterface(t *testing.T) {
 	for _, generifyInterface := range [...]bool{true, false} {
 		interfaceTypeResult := TypeNameByReflect[testInterface](generifyInterface)
 		pointerTypeResult := TypeNameByReflect[*testInterface](generifyInterface)
+		doublePointerTypeResult := TypeNameByReflect[**testInterface](generifyInterface)
 		if interfaceTypeResult == pointerTypeResult {
-			t.Error("type names was the same, when it should be different, because of different pointer")
+			t.Error(typeNameErr)
+		}
+		if doublePointerTypeResult == pointerTypeResult {
+			t.Error(typeNameErr)
 		}
 	}
 }