+++ remy/remy.go
@@ -2,6 +2,7 @@ package remy
 
 import (
 	"errors"
+	"fmt"
 	"github.com/wrapped-owls/goremy-di/remy/internal/injector"
 	"github.com/wrapped-owls/goremy-di/remy/internal/types"
 	"github.com/wrapped-owls/goremy-di/remy/internal/utils"
@@ -14,7 +15,7 @@ type (
 
 	// Bind is directly copy from types.Bind
 	Bind[T any] interface {
-		Generates(DependencyRetriever) T
+		types.Bind[T]
 	}
 
 	// Config defines needed configuration to instantiate a new injector
@@ -61,11 +62,8 @@ func NewInjector(configs ...Config) Injector {
 
 // Register must be called first, because the library doesn't support registering dependencies while get at same time.
 // This is not supported in multithreading applications because it does not have race protection
-func Register[T any](i Injector, bind types.Bind[T], keys ...string) {
-	if i == nil {
-		i = globalInjector()
-	}
-	if err := injector.Register(i, bind, keys...); err != nil {
+func Register[T any](i Injector, bind Bind[T], keys ...string) {
+	if err := injector.Register[T](mustInjector(i), bind, keys...); err != nil {
 		panic(err)
 	}
 }
@@ -75,11 +73,10 @@ func Register[T any](i Injector, bind types.Bind[T], keys ...string) {
 // the library doesn't support registering dependencies while get at same time.
 //
 // This is not supported in multithreading applications because it does not have race protection
-func Override[T any](i Injector, bind types.Bind[T], keys ...string) {
-	if i == nil {
-		i = globalInjector()
-	}
-	if err := injector.Register(i, bind, keys...); err != nil && !errors.Is(err, utils.ErrAlreadyBound) {
+func Override[T any](i Injector, bind Bind[T], keys ...string) {
+	if err := injector.Register[T](mustInjector(i), bind, keys...); err != nil && !errors.Is(
+		err, utils.ErrAlreadyBound,
+	) {
 		panic(err)
 	}
 }
@@ -89,10 +86,12 @@ func Override[T any](i Injector, bind types.Bind[T], keys ...string) {
 // Receives: Injector (required); value (required); key (optional)
 func RegisterInstance[T any](i Injector, value T, keys ...string) {
 	Register(
-		i,
-		Instance(func(DependencyRetriever) T {
-			return value
-		}),
+		mustInjector(i),
+		Instance(
+			func(DependencyRetriever) T {
+				return value
+			},
+		),
 		keys...,
 	)
 }
@@ -102,10 +101,12 @@ func RegisterInstance[T any](i Injector, value T, keys ...string) {
 // Receives: Injector (required); value (required); key (optional)
 func RegisterSingleton[T any](i Injector, value T, keys ...string) {
 	Register(
-		i,
-		Singleton(func(DependencyRetriever) T {
-			return value
-		}),
+		mustInjector(i),
+		Singleton(
+			func(DependencyRetriever) T {
+				return value
+			},
+		),
 		keys...,
 	)
 }
@@ -114,61 +115,64 @@ func RegisterSingleton[T any](i Injector, value T, keys ...string) {
 //
 // Receives: DependencyRetriever (required); key (optional)
 func Get[T any](i DependencyRetriever, keys ...string) T {
-	if i == nil {
-		i = globalInjector()
-	}
-	return injector.TryGet[T](i, keys...)
+	return injector.TryGet[T](mustRetriever(i), keys...)
 }
 
 // DoGet directly access a retriever and returns the type that was bound in it.
 // Additionally, it returns an error which indicates if the bind was found or not.
 //
 // Receives: DependencyRetriever (required); key (optional)
-func DoGet[T any](i DependencyRetriever, keys ...string) (T, error) {
-	if i == nil {
-		i = globalInjector()
-	}
-	return injector.Get[T](i, keys...)
+func DoGet[T any](i DependencyRetriever, keys ...string) (result T, err error) {
+	defer func() {
+		r := recover()
+		if r != nil {
+			err = fmt.Errorf("%v", r)
+		}
+	}()
+
+	return injector.Get[T](mustRetriever(i), keys...)
 }
 
 // GetGen creates a sub-injector and access the retriever to generate and return a Factory bind
 //
-// Receives: Injector (required); []InstancePairAny (required); key (optional)
-func GetGen[T any](ij Injector, elements []InstancePairAny, keys ...string) T {
-	if ij == nil {
-		ij = globalInjector()
-	}
-	return injector.TryGetGen[T](ij, elements, keys...)
+// Receives: DependencyRetriever (required); []InstancePairAny (required); key (optional)
+func GetGen[T any](i DependencyRetriever, elements []InstancePairAny, keys ...string) T {
+	return injector.TryGetGen[T](mustRetriever(i), elements, keys...)
 }
 
 // DoGetGen creates a sub-injector and access the retriever to generate and return a Factory bind
 // Additionally, it returns an error which indicates if the bind was found or not.
 //
-// Receives: Injector (required); []InstancePairAny (required); key (optional)
-func DoGetGen[T any](ij Injector, elements []InstancePairAny, keys ...string) (T, error) {
-	if ij == nil {
-		ij = globalInjector()
-	}
-	return injector.GetGen[T](ij, elements, keys...)
+// Receives: DependencyRetriever (required); []InstancePairAny (required); key (optional)
+func DoGetGen[T any](i DependencyRetriever, elements []InstancePairAny, keys ...string) (result T, err error) {
+	defer func() {
+		r := recover()
+		if r != nil {
+			err = fmt.Errorf("%v", r)
+		}
+	}()
+
+	return injector.GetGen[T](mustRetriever(i), elements, keys...)
 }
 
 // GetGenFunc creates a sub-injector and access the retriever to generate and return a Factory bind
 //
-// Receives: Injector (required); func(Injector) (required); key (optional)
-func GetGenFunc[T any](ij types.Injector, binder func(Injector), keys ...string) T {
-	if ij == nil {
-		ij = globalInjector()
-	}
-	return injector.TryGetGenFunc[T](ij, binder, keys...)
+// Receives: DependencyRetriever (required); func(Injector) (required); key (optional)
+func GetGenFunc[T any](i DependencyRetriever, binder func(Injector), keys ...string) T {
+	return injector.TryGetGenFunc[T](mustRetriever(i), binder, keys...)
 }
 
 // DoGetGenFunc creates a sub-injector and access the retriever to generate and return a Factory bind
 // Additionally, it returns an error which indicates if the bind was found or not.
 //
-// Receives: Injector (required); func(Injector) (required); key (optional)
-func DoGetGenFunc[T any](ij types.Injector, binder func(Injector), keys ...string) (T, error) {
-	if ij == nil {
-		ij = globalInjector()
-	}
-	return injector.GetGenFunc[T](ij, binder, keys...)
+// Receives: DependencyRetriever (required); func(Injector) (required); key (optional)
+func DoGetGenFunc[T any](i DependencyRetriever, binder func(Injector), keys ...string) (result T, err error) {
+	defer func() {
+		r := recover()
+		if r != nil {
+			err = fmt.Errorf("%v", r)
+		}
+	}()
+
+	return injector.GetGenFunc[T](mustRetriever(i), binder, keys...)
 }